refactor(api): Move artifact deployment events to keel-api
