--- keel-core/src/main/kotlin/com/netflix/spinnaker/keel/events/ArtifactVersionDeploying.kt
@@ -1,9 +1,0 @@
-package com.netflix.spinnaker.keel.events
-
-/**
- * An event fired when we launch a task to deploy a new version to a cluster
- */
-data class ArtifactVersionDeploying(
-  val resourceId: String,
-  val artifactVersion: String
-)
